use size box instead of padding and add all missing trailing commas
