+++ lib/assesments/assessment_2/fi-as2-hira/Screens/Login/fi_as2_hira_login.dart
@@ -1,6 +1,7 @@
 import 'package:dotted_border/dotted_border.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_interns/assesments/assessment_2/fi-as2-hira/constants/fi_as2_hira_colors.dart';
+import 'package:flutter_interns/assesments/assessment_2/fi-as2-hira/constants/fi_as2_hira_spaces.dart';
 import '../../widgets/core/fi_as2_hira_button.dart';
 import '../../widgets/core/fi_as2_hira_home_root.dart';
 part 'widgets/fi_as2_hira_dottedborder.dart';
@@ -12,9 +13,7 @@ class H1Login extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    Widget smallspace = const SizedBox(
-      height: 35,
-    );
+   
     return Scaffold(
       backgroundColor: Colors.white,
       body: SafeArea(
@@ -27,7 +26,7 @@ class H1Login extends StatelessWidget {
             const Center(
               child: H1Logo(),
             ),
-            smallspace,
+            Spaces.h35,
             const Text(
               "Let's Connect \n    Together",
               style: TextStyle(
@@ -36,7 +35,7 @@ class H1Login extends StatelessWidget {
                 color: Constants.black,
               ),
             ),
-            smallspace,
+            Spaces.h35,
             H1Button(
               onPressed: () {
                 Navigator.push(
@@ -56,9 +55,7 @@ class H1Login extends StatelessWidget {
                 ),
               ),
             ),
-            const SizedBox(
-              height: 15,
-            ),
+            Spaces.h15,
             H1Button(
               onPressed: () {
                 Navigator.push(
