+++ lib/assesments/assessment_2/fi-as2-hira/Screens/Login/widgets/fi_as2_hira_circlecontainer.dart
@@ -10,7 +10,7 @@ class H1Circle extends StatelessWidget {
       height: 50,
       width: 50,
       decoration: BoxDecoration(
-          color: Constants.grey, borderRadius: BorderRadius.circular(50)),
+          color: Constants.grey, borderRadius: BorderRadius.circular(50),),
       padding: const EdgeInsets.all(3.0),
       child: ClipRRect(
         borderRadius: BorderRadius.circular(50),
