+++ lib/assesments/assessment_2/fi-as2-hira/Screens/Login/widgets/fi_as2_hira_dottedborder.dart
@@ -14,10 +14,11 @@ class H1Dotted extends StatelessWidget {
       dashPattern: const [11, 9],
       padding: const EdgeInsets.all(6),
       child: ClipRRect(
-          borderRadius: const BorderRadius.all(
-            Radius.circular(500),
-          ),
-          child: child),
+        borderRadius: const BorderRadius.all(
+          Radius.circular(500),
+        ),
+        child: child,
+      ),
     );
   }
 }
