+++ lib/assesments/assessment_2/fi-as2-hira/Screens/my_profile/widgets/fi_as2_hira_text.dart
@@ -12,17 +12,19 @@ class H1Text extends StatelessWidget {
         Text(
           text1,
           style: const TextStyle(
-              fontSize: 16,
-              fontWeight: FontWeight.w600,
-              color: Constants.fontgrey),
+            fontSize: 16,
+            fontWeight: FontWeight.w600,
+            color: Constants.fontgrey,
+          ),
         ),
         Text(
           text2,
           style: const TextStyle(
-              fontSize: 20,
-              fontWeight: FontWeight.w600,
-              color: Constants.black),
-        )
+            fontSize: 20,
+            fontWeight: FontWeight.w600,
+            color: Constants.black,
+          ),
+        ),
       ],
     );
   }
