--- lib/assesments/assessment_2/fi-as2-hira/widgets/design/fi_as2_hira_addbox.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/material.dart';
-
-class H1Addbox extends StatelessWidget {
-  const H1Addbox({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return  Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        elevation: 0.0,
-        leading: IconButton(
-          icon: const Icon(
-            Icons.arrow_back,
-            color: Colors.black,
-          ),
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
-      ),
-      body: const Center(
-          child: Text(
-        'Hi Go back!\n AddBox ',
-        style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
-      )),
-    );
-  }
-}
